gneb_workflow: added support for a scratch_folder for the running updates of the chain file, which is separate from the output folder
The main use is when running on a cluster with a synchronizing filesystem, where the saves of the chain file after each step can now be done on the local disk (of the node in the scratch folder) , while the rest of the output remains in the synchronized filesystem. In the scratch folder a new folder with a unique name gets created. After the run() function finishes the chain gets copied to the output folder automatically.